small cleanup for path traversal
- use get_result(), singular, to get single result
- remove dead `max_targets` param from cypher rendering
- move ReachableNodesExecutor to executor.py
